* origin/master:
  ELEC-261: Actually use GCC's warnings (#74)

+++ libraries/STM32F0xx_StdPeriph_Driver/src/stm32f0xx_interrupt.c
@@ -52,7 +52,7 @@ StatusCode stm32f0xx_interrupt_exti_enable(uint8_t line, const InterruptSettings
     return status_code(STATUS_CODE_INVALID_ARGS);
   }
 
-  EXTI_InitTypeDef init_struct = { .EXTI_Line = 0x01 << line,
+  EXTI_InitTypeDef init_struct = { .EXTI_Line = (uint32_t)0x01 << line,
                                    .EXTI_Mode = 0x04 * settings->type,
                                    .EXTI_Trigger = 0x08 + 0x04 * edge,
                                    .EXTI_LineCmd = ENABLE };
@@ -67,7 +67,7 @@ StatusCode stm32f0xx_interrupt_exti_trigger(uint8_t line) {
     return status_code(STATUS_CODE_INVALID_ARGS);
   }
 
-  EXTI_GenerateSWInterrupt(0x01 << line);
+  EXTI_GenerateSWInterrupt((uint32_t)0x01 << line);
   return STATUS_CODE_OK;
 }
 
@@ -76,7 +76,7 @@ StatusCode stm32f0xx_interrupt_exti_get_pending(uint8_t line, uint8_t *pending_b
     return status_code(STATUS_CODE_INVALID_ARGS);
   }
 
-  *pending_bit = (uint8_t)EXTI_GetITStatus(0x01 << line);
+  *pending_bit = (uint8_t)EXTI_GetITStatus((uint32_t)0x01 << line);
   return STATUS_CODE_OK;
 }
 
@@ -85,6 +85,6 @@ StatusCode stm32f0xx_interrupt_exti_clear_pending(uint8_t line) {
     return status_code(STATUS_CODE_INVALID_ARGS);
   }
 
-  EXTI_ClearITPendingBit(0x01 << line);
+  EXTI_ClearITPendingBit((uint32_t)0x01 << line);
   return STATUS_CODE_OK;
 }
