+++ libraries/ms-common/test/test_pqueue_backed.c
@@ -20,13 +20,13 @@ void teardown_test(void) {}
 void test_pqueue_backed_run(void) {
   uint16_t prios[] = { 50, 10, 20, 2, 17, 5, 4000, 0, 3, 240 };
 
-  for (int i = 0; i < SIZEOF_ARRAY(prios); i++) {
+  for (size_t i = 0; i < SIZEOF_ARRAY(prios); i++) {
     StatusCode result = pqueue_backed_push(&s_queue, &(TestObject){.data = prios[i] }, prios[i]);
     TEST_ASSERT_EQUAL(STATUS_CODE_OK, result);
   }
 
   uint16_t last_prio = 0;
-  for (int i = 0; i < SIZEOF_ARRAY(prios); i++) {
+  for (size_t i = 0; i < SIZEOF_ARRAY(prios); i++) {
     TestObject obj;
     StatusCode result = pqueue_backed_pop(&s_queue, &obj);
     TEST_ASSERT_EQUAL(STATUS_CODE_OK, result);
